Create epic game store and ebay functions
Created the functions that grab the price info for epic games store and ebay. Ebay function can grab the title and price need to add the shipping and format (auction, buy it now, or best offer). Epic Game store does not give html to take needed info need to find alternative way to grab it.