Updated the Ebay function and Project roadmap
Fixed error in the ebay function not sending the title of the first item along with problems associated with different types of formats for the shipping on different listings. Restructured the function making it much more efficient (about 20 lines of code shorter!). Also fixed an error with the Steam function failing due to not finding a summary text for certain listings.